kafka-consumer-pg/requirements.txt
kafka-python
psycopg2
python-dotenv
fastapi
uvicorn
